# openapi.yaml
openapi: 3.0.3
info:
  title: Coachwise API
  version: "0.1.0"
  description: >
    Early-stage REST API for Coachwise. All non-auth routes require
    `Authorization: Bearer <access_token>`.
servers:
  - url: http://localhost:3000
paths:
...
  /auth/pre-register:
    post:
      summary: Check email/username availability
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                username:
                  type: string
      responses:
        "200":
          description: Availability status
          content:
            application/json:
              schema:
                type: object
                properties:
                  email:
                    type: string
                    example: AVAILABLE
                  username:
                    type: string
                    example: EXISTS
  /auth/register:
    post:
      summary: Register user (email/password)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterForm'
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  id:
                    type: string
                    format: uuid
  /auth/login:
    post:
      summary: Login with email or username
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginForm'
      responses:
        "200":
          description: Tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
  /auth/refresh:
    post:
      summary: Refresh tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
      responses:
        "200":
          description: New tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
  /auth/logout:
    post:
      summary: Logout and blacklist access token
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Success
  /auth/otp:
    post:
      summary: Send OTP to email (auth)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
      responses:
        "200":
          description: Success
  /auth/otp/verify:
    post:
      summary: Verify OTP and issue tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                code:
                  oneOf:
                    - type: string
                    - type: integer
      responses:
        "200":
          description: Tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
  /auth/password/forget:
    post:
      summary: Send OTP for password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
      responses:
        "200":
          description: Success
  /auth/password:
    put:
      summary: Change password (normal or direct)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/DirectPasswordChangeForm'
                - $ref: '#/components/schemas/NormalPasswordChangeForm'
      responses:
        "202":
          description: Password updated
...
  /exercises:
    get:
      summary: List exercises
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: public
          schema:
            type: boolean
        - in: query
          name: name
          schema:
            type: string
      responses:
        "200":
          description: Exercises
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Exercise'
    post:
      summary: Create exercise
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExerciseForm'
      responses:
        "201":
          description: Created exercise
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exercise'
  /exercises/{id}:
    get:
      summary: Get exercise
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Exercise
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exercise'
    put:
      summary: Update exercise
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExerciseForm'
      responses:
        "200":
          description: Updated exercise
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exercise'
    delete:
      summary: Delete exercise
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AuthTokens:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token:
          type: string
    RegisterForm:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        username:
          type: string
        first_name:
          type: string
        last_name:
          type: string
    LoginForm:
      type: object
      properties:
        email:
          type: string
          format: email
        username:
          type: string
        password:
          type: string
      required: [password]
    DirectPasswordChangeForm:
      type: object
      properties:
        password:
          type: string
      required: [password]
    NormalPasswordChangeForm:
      type: object
      properties:
        current_password:
          type: string
        password:
          type: string
      required: [current_password, password]
...
    Exercise:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
        description:
          type: string
        public:
          type: boolean
        sets:
          type: array
          items:
            $ref: '#/components/schemas/ExerciseSet'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ExerciseSet:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          nullable: true
        exercise_id:
          type: string
          format: uuid
        set_number:
          type: integer
        rest_time:
          type: integer
          description: nanoseconds
        rep_count:
          type: integer
          nullable: true
        duration:
          type: integer
          description: nanoseconds
          nullable: true
    ExerciseForm:
      type: object
      required: [name, description]
      properties:
        name:
          type: string
        description:
          type: string
        public:
          type: boolean
        sets:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              rest_time:
                type: integer
              rep_count:
                type: integer
                nullable: true
              duration:
                type: integer
                nullable: true
...